Refactor ui package
Move Ui class into ui package.
